CS60003/hw2/task3/requirements.txt
torch
torchvision
numpy
Pillow
PyYAML
matplotlib
swanlab
modelscope
